+++ test/e2e/scripts/copy.js
@@ -2,11 +2,8 @@
 
 const fs = require('fs');
 
-function copy(dir) {
-    fs.copyFileSync(
-        './test/e2e/assets/script.js',
-        `./test/e2e/generated/${dir}/script.js`,
-    );
+function copy(fileNameIn, fileNameOut) {
+    fs.copyFileSync(`./test/e2e/assets/${fileNameIn}`, `./test/e2e/generated/${fileNameOut}`);
 }
 
 module.exports = copy;