+++ test/e2e/scripts/server.js
@@ -1,9 +1,10 @@
 'use strict';
 
 const express = require('express');
+const path = require('path');
 
 let app;
-let server
+let server;
 
 async function start(dir) {
     return new Promise(resolve => {
@@ -12,32 +13,41 @@ async function start(dir) {
         // Serve the JavaScript files from the specific folder, since we are using browser
         // based ES6 modules, this also means that we can just request the js/index.js file
         // and all other relative paths are resolved from that file.
-        app.use('/js', express.static(`./test/e2e/generated/${dir}/`, {
-            extensions: ['', 'js'],
-            index: 'index.js',
-        }));
+        app.use(
+            '/js',
+            express.static(`./test/e2e/generated/${dir}/`, {
+                extensions: ['', 'js'],
+                index: 'index.js',
+            })
+        );
 
         // When we request the index then we can just return the script loader.
         // This file is copied from test/e2e/assets/script.js to the output directory
         // of the specific version and client.
-        app.get('/', (req, res) => {
-            res.send('<script src="js/script.js"></script>');
+        app.get('/', (_req, res) => {
+            res.sendFile(path.resolve(`./test/e2e/generated/${dir}/index.html`));
+        });
+
+        app.get('/script.js', (_req, res) => {
+            res.sendFile(path.resolve(`./test/e2e/generated/${dir}/script.js`));
         });
 
         // Register an 'echo' server that just returns all data from the API calls.
         // Although this might not be a 'correct' response, we can use this to test
         // the majority of API calls.
         app.all('/base/api/*', (req, res) => {
-            res.json({
-                method: req.method,
-                protocol: req.protocol,
-                hostname: req.hostname,
-                path: req.path,
-                url: req.url,
-                query: req.query,
-                body: req.body,
-                headers: req.headers,
-            });
+            setTimeout(() => {
+                res.json({
+                    method: req.method,
+                    protocol: req.protocol,
+                    hostname: req.hostname,
+                    path: req.path,
+                    url: req.url,
+                    query: req.query,
+                    body: req.body,
+                    headers: req.headers,
+                });
+            }, 100);
         });
 
         server = app.listen(3000, resolve);
@@ -45,8 +55,12 @@ async function start(dir) {
 }
 
 async function stop() {
-    return new Promise(resolve => {
-        server.close(resolve);
+    return new Promise((resolve, reject) => {
+        try {
+            server.close(resolve);
+        } catch (err) {
+            reject(err);
+        }
     });
 }