+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.7] - 2022-10-01
+
+### Added
+
+- Added cancelable concept
+
+### Fixed
+
+- Fixed incorrect type of error in catchErrors function
+- Fixed node version in CI
+- Fixed dependencies
+
 ## [0.2.6] - 2022-09-18
 
 ### Added